move sampler under opto/features/priority_search
