+++ opto/features/priority_search/utils.py
@@ -9,7 +9,7 @@ from opto.trainer.utils import async_run, batch_run
 from opto.optimizers.utils import print_color
 from opto.trainer.algorithms.basic_algorithms import Minibatch, AlgorithmBase, batchify
 from opto.trainer.loader import DataLoader
-from opto.trainer.sampler import Sampler, RolloutsGraph
+from opto.features.priority_search.sampler import Sampler, RolloutsGraph
 import time
 
 # Some helper functions to convert between trace.Module and update_dict
